Interface

SingleCellProjections.DataMatrix

SingleCellProjections.DataMatrix

SingleCellProjections.DataMatrix

SingleCellProjections.LowRank

SingleCellProjections.NormalizationModel

SingleCellProjections.SCTransformModel

SingleCellProjections.SVDModel

LinearAlgebra.svd

SCTransform.sctransform

SingleCellProjections.covariate

SingleCellProjections.designmatrix

SingleCellProjections.differentialexpression

SingleCellProjections.filter_matrix

SingleCellProjections.filter_obs

SingleCellProjections.filter_var

SingleCellProjections.force_layout

SingleCellProjections.implicitsvd

SingleCellProjections.load10x

SingleCellProjections.load_counts

SingleCellProjections.load_counts

SingleCellProjections.load_counts

SingleCellProjections.loadh5ad

SingleCellProjections.logtransform

SingleCellProjections.merge_counts

SingleCellProjections.normalize_matrix

SingleCellProjections.normalize_matrix

SingleCellProjections.obs_coordinates

SingleCellProjections.project

SingleCellProjections.project

SingleCellProjections.project

SingleCellProjections.splitrange

SingleCellProjections.tf_idf_transform

SingleCellProjections.update_matrix

SingleCellProjections.var_coordinates

SingleCellProjections.var_counts_fraction!

SingleCellProjections.variable_std

SingleCellProjections.variable_var

SingleCellProjections.DataMatrix — Type

struct DataMatrix{T,Tv,To}

A DataMatrix represents a matrix together with annotations for variables and observations.

Fields:

matrix::T - The matrix.

var::Tv - Variable annotations.

obs::To - Observation annotations.

var_id_cols::Vector{String} - Which column(s) to use as IDs.

obs_id_cols::Vector{String} - Which column(s) to use as IDs.

models::Vector{ProjectionModel} - Models used in the creation of this DataMatrix.

The rows of the var and obs tables must be unique, considering only the var_id_cols/obs_id_cols.

SingleCellProjections.DataMatrix — Method

DataMatrix(matrix, var, obs; var_id_cols=nothing, obs_id_cols=nothing)

Create a DataMatrix with the given matrix, var and obs.

Columns to use for var/obs IDs can be explicitly set with var_id_cols/obs_id_cols. Otherwise, an attempt will be made to autodetect the ID columns.

SingleCellProjections.DataMatrix — Method

DataMatrix()

Create an empty DataMatrix{Matrix{Float64},DataFrame,DataFrame}.

SingleCellProjections.LowRank — Type

LowRank

A matrix decomposition UVᵀ where each row of U represents a variable and each column of Vᵀ represents a sample. Intended for situations where the product is low rank, i.e. size(U,2)==size(Vt,1) is small.

SingleCellProjections.NormalizationModel — Method

NormalizationModel(data::DataMatrix, design::DesignMatrix;
                   scale=false, min_std=1e-6, annotate=true,
                   rtol=sqrt(eps()), var=:copy, obs=:copy)

Create a NormalizationModel based on data and a design matrix.

scale - Set to true to normalize variables to unit standard deviation. Can also be set to a vector with a scaling factor for each variable.

min_std - If scale==true, the scale vector is set to 1.0 ./ max.(std, min_std). That is, min_std is used to suppress variables that are very small (and any fluctuations can be assumed to be noise).

annotate - Only used if scale!=false. With annotate=true, the scale vector is added as a var annotation.

rtol - Singular values of the design matrix that are ≤rtol are discarded. Needed for numerical stability.

var - Can be :copy (make a copy of source var) or :keep (share the source var object).

obs - Can be :copy (make a copy of source obs) or :keep (share the source obs object).

See also: normalize_matrix, designmatrix

SingleCellProjections.SCTransformModel — Method

SCTransformModel(counts::DataMatrix;
                 var_filter = hasproperty(counts.var, :feature_type) ? :feature_type => isequal("Gene Expression") : nothing,
                 rtol=1e-3, atol=0.0, annotate=true,
                 post_var_filter=:, post_obs_filter=:,
                 obs=:copy,
                 kwargs...)

Computes the SCTransform parameter estimates for counts and creates a SCTransformModel that can be applied to the same or another data set. Defaults to only using "Gene Expression" features.

var_filter - Control which variables (features) to use for parameter estimation. Defaults to :feature_type => isequal("Gene Expression"), if a feature_type column is present in counts.var. Can be set to nothing to disable filtering. See DataFrames.filter for how to specify filters.

rtol - Relative tolerance when constructing low rank approximation.

atol - Absolute tolerance when constructing low rank approximation.

annotate - Set to true to include SCTransform parameter estimates as feature annotations.

post_var_filter - Equivalent to applying variable (feature) filtering after sctransform, but computationally more efficient.

post_obs_filter - Equivalent to applying observation (cell) filtering after sctransform, but computationally more efficient.

obs - Can be :copy (make a copy of source obs) or :keep (share the source obs object).

kwargs... - Additional kwargs are passed on to SCTransform.scparams.

Examples

Setup SCTransformModel (Gene Expression features):

julia> SCTransformModel(counts)

Setup SCTransformModel (Antibody Capture features):

julia> SCTransformModel(counts; var_filter = :feature_type => isequal("Antibody Capture"))

See also: sctransform, SCTransform.scparams, DataFrames.filter

SingleCellProjections.SVDModel — Type

SVDModel <: ProjectionModel

A model used for projecting onto an SVD object. Normally created using svd(::DataMatrix).

See also: svd

LinearAlgebra.svd — Method

svd(data::DataMatrix; nsv=3, var=:copy, obs=:copy, kwargs...)

Compute the Singular Value Decomposition (SVD) of data using the Random Subspace SVD algorithm from [Halko et al. "Finding structure with randomness: Probabilistic algorithms for constructing approximate matrix decompositions"]. SVD is often used to perform Principal Component Analysis (PCA), which assumes that the data is centered.

nsv - The number of singular values.

var - Can be :copy (make a copy of source var) or :keep (share the source var object).

obs - Can be :copy (make a copy of source obs) or :keep (share the source obs object).

Additional kwargs related to numerical precision are passed to SingleCellProjections.implicitsvd.

See also: SingleCellProjections.implicitsvd

SCTransform.sctransform — Method

sctransform(counts::DataMatrix; verbose=true, kwargs...)

Compute the SCTransform of the DataMatrix counts. The result is stored as a Matrix Expression with the sum of a sparse and a low-rank term. I.e. no large dense matrix is created.

See SCTransformModel for description of kwargs....

Examples

Compute SCTransform (Gene Expression features):

julia> sctransform(counts)

Compute SCTransform (Antibody Capture features):

julia> sctransform(counts; var_filter = :feature_type => isequal("Antibody Capture"))

See also: SCTransformModel, SCTransform.scparams

SingleCellProjections.covariate — Function

covariate(name::String, type=:auto)

Create a covariate referring to column name. type must be one of :auto, :numerical, :categorical and :intercept. :auto means auto-detection by checking if the values in the column are numerical or categorical. type==:intercept adds an intercept to the model (in which case the name parameter is ignored).

See also: designmatrix

SingleCellProjections.designmatrix — Method

designmatrix(data::DataMatrix, [covariates...]; center=true, max_categories=100)

Creates a design matrix from data.obs and the given covariates. Covariates can be specied using strings (column name in data.obs), with autodetection of whether the covariate is numerical or categorical, or using the covariate function for more control.

center - If true, an intercept is added to the design matrix. (Should only be set to false in very rare circumstances.)

max_categories - Safety parameter, an error will be thrown if there are too many categories. In this case, it is likely a mistake that the covariate was used as a categorical covariate. Using a very large number of categories is also bad for performance and memory consumption.

Examples

Centering only:

julia> designmatrix(data)

Regression model with intercept (centering) and "fraction_mt" (numerical annotation):

julia> designmatrix(data, "fraction_mt")

As above, but also including "batch" (categorical annotation):

julia> designmatrix(data, "fraction_mt", "batch")

See also: normalize_matrix, NormalizationModel, covariate

SingleCellProjections.differentialexpression — Method

differentialexpression(data::DataMatrix{<:SparseMatrixCSC}, groups::AbstractVector)

Assumes data is logtransformed. groups should be a vector with 1 for observations in group 1, 2 for observations in group 2 and missing for other observations.

SingleCellProjections.filter_matrix — Method

filter_matrix(fvar, fobs, data::DataMatrix)

Return a new DataMatrix, containing only the variables and observations passing the filters.

fvar/fobs can be:

An AbstractVector of indices to keep.

A AbstractVector of booleans (true to keep, false to discard).

: indicating that all variables/observations should be kept.

Anything you can pass on to DataFrames.filter (see DataFrames documentation for details).

Also note that indexing of a DataMatrix supports AbstractVectors of indices/booleans and :, and is otherwise identical to filter_matrix.

Examples

Keep every 10th variable and 3rd observation:

julia> filter_matrix(1:10:size(data,1), 1:3:size(data,2), data)

Or, using indexing syntax:

julia> data[1:10:end, 1:3:end]

For more examples, see filter_var and filter_obs.

See also: filter_var, filter_obs, DataFrames.filter

SingleCellProjections.filter_obs — Method

filter_obs(f, data::DataMatrix)

Return a new DataMatrix, containing only the observations passing the filter.

f can be:

An AbstractVector of indices to keep.

A AbstractVector of booleans (true to keep, false to discard).

: indicating that all observations should be kept.

Anything you can pass on to DataFrames.filter (see DataFrames documentation for details).

Examples

Keep every 10th observation:

julia> filter_obs(1:10:size(data,2), data)

Remove observations where "celltype" equals "other":

julia> filter_obs("celltype"=>!isequal("other"), data)

See also: filter_matrix, filter_var, DataFrames.filter

SingleCellProjections.filter_var — Method

filter_var(f, data::DataMatrix)

Return a new DataMatrix, containing only the variables passing the filter.

f can be:

An AbstractVector of indices to keep.

A AbstractVector of booleans (true to keep, false to discard).

: indicating that all variables should be kept.

Anything you can pass on to DataFrames.filter (see DataFrames documentation for details).

Examples

Keep every 10th variable:

julia> filter_var(1:10:size(data,1), data)

Keep only variables of the type "Gene Expression":

julia> filter_var("feature_type"=>isequal("Gene Expression"), data)

See also: filter_matrix, filter_obs, DataFrames.filter

SingleCellProjections.force_layout — Method

force_layout(data::DataMatrix;
             ndim=3,
             k,
             adj,
             kprojection=10,
             obs=:copy,
             adj_out,
             niter = 100,
             link_distance = 4,
             link_strength = 2,
             charge = 5,
             charge_min_distance = 1,
             theta = 0.9,
             center_strength = 0.05,
             velocity_decay = 0.9,
             initialAlpha = 1.0,
             finalAlpha = 1e-3,
             initialScale = 10,
             rng)

Compute the Force Layout (also known as a force directed knn-graph or SPRING plots) for data. Usually, data is a DataMatrix after reduction to 10-100 dimensions by svd.

A Force Layout is computed by running a physics simulation were the observations are connected by springs (such that connected observations are attracted), a general "charge" force repelling all observations from each other and a centering force that keep the observations around the origin. The implementation is based on d3-force: https://github.com/d3/d3-force, also see LICENSE.md.

Exactly one of the kwargs k and adj must be provided. See details below.

General parameters:

k - Number of nearest neighbors to connect each observation to (computes adj below).

adj - An sparse, symmetric, adjacency matrix with booleans. true if two observations are connected by a spring and false otherwise.

kprojection - The number of nearest neighbors used when projecting onto the resulting force layout. (Not used in the computation of the layout, only during projection.)

obs - Can be :copy (make a copy of source obs) or :keep (share the source obs object).

adj_out - Optional Ref. If specified, the (computed) adj matrix will be assigned to adj_out.

Paramters controlling the physics simulation:

niter - Number of iterations to run the simulation.

link_distance - The length of each spring.

link_strength - The strength of the spring force.

charge - The strength of the charge force.

charge_min_distance - Used to avoid numerical instabilities by limiting the charge force for observations that are very close.

theta - Parameter controlling accuracy in the Barnes-Hut approximation for charge forces.

center_strength - Strength of the centering force.

velocity_decay - At each iteration, the current velocity for an observations is multiplied by velocity_decay.

initialAlpha - The alpha value decreases over time and allows larger changes to happen early, while being more stable towards the end.

finalAlpha - See initialAlpha

initialScale - The simulation is initialized by randomly drawing each observation from a multivariate Gaussian, and is scaled by initialScale.

rng - Optional RNG object. Useful for reproducibility.

Examples

julia> force_layout(data; ndim=3, k=100)

SingleCellProjections.implicitsvd — Method

implicitsvd(A; nsv=3, subspacedims=4nsv, niter=2, stabilize_sign=true, rng)

Compute the SVD of A using Random Subspace SVD. [Halko et al. "Finding structure with randomness: Probabilistic algorithms for constructing approximate matrix decompositions"]

nsv - Number of singular values/vectors to compute

subspacedims - Number of dimensions used for the subspace approximating the action of A.

niter - Number of iterations. In each iteration, one multiplication of A with a matrix and one multiplication of A' with a matrix will be performed.

stabilize_sign - If true, handles the problem that the SVD is only unique up to the sign of each component (for real matrices), by ensuring that the l1 norm of the positive entires for each column in U is larger than the l1 norm of the negative entries.

rng - Specify a custom RNG.

SingleCellProjections.load10x — Method

load10x(filename; lazy=false, copy_obs_col="barcode"=>"id", kwargs...)

Load a CellRanger ".h5" or ".mtx[.gz]" file as a DataMatrix.

lazy - If true, the count matrix itself will not be loaded, only features and barcodes. This is used internally in load_counts to merge samples more efficiently. Use load_counts to later load the count data.

copy_obs_col - Set to nothing to disable. Defaults to a pair "barcode"=>"id", which copies the obs annotation barcode to a new column id.

Additional kwargs... are passed to the DataMatrix constructor.

Examples

Load counts from a CellRanger ".h5" file. (Recommended.)

julia> counts = load10x("filtered_feature_bc_matrix.h5")

Load counts from a CellRanger ".mtx" file. Tries to find barcode and feature annotation files in the same folder.

julia> counts = load10x("matrix.mtx.gz")

Lazy loading followed by loading.

julia> counts = load10x("filtered_feature_bc_matrix.h5");
julia> counts = load_counts(counts)

See also: load_counts

SingleCellProjections.load_counts — Method

load_counts([loadfun=load10x], filenames;
            sample_names,
            sample_name_col,
            merged_obs_id_col = "id",
            lazy,
            lazy_merge = false,
            var_id_cols=nothing,
            merged_obs_id_delim = '_',
            callback=nothing)

Load and merge multiple samples efficiently.

Defaults to loading 10x CellRanger files. The files are first loaded lazily, then the merged count matrix is allocated and finally each sample is loaded directly into the merged count matrix. (This strategy greatly reduces memory usage, since only one copy of data is needed instead of two.)

The vector filenames specifies which files to load. For each file, loadfun is called.

sample_names - Specify the sample names. Should be a vector of the same length as filenames. Set to nothing to not create a sample name annotation.

sample_name_col - Column for sample names in obs, defaults to "sampleName".

merged_obs_id_col - Colum for merged ids in obs.

lazy - Enable lazy loading. Defaults to true if load10x is used, and false otherwise.

lazy_merge - Enable lazy merging, i.e. var and obs are created, but the count matrix merging is postponed until a second call to load_counts.

var_id_cols - Specify variable id columns. Defaults to nothing, i.e. autodetecting from the samples.

merged_obs_id_delim - Delimiter used when creating merged obs IDs.

callback - Experimental callback functionality. The callback function is called between samples during merging. Return true to abort loading and false to continue.

Examples

Load and name samples:

julia> counts = load_counts(["s1.h5", "s2.h5"]; sample_names=["Sample A", "Sample B"])

See also: load10x, merge_counts

SingleCellProjections.load_counts — Method

load_counts(data::DataMatrix{<:Lazy10xMatrix})

Load counts for a lazily loaded 10x DataMatrix.

See also: load10x

SingleCellProjections.load_counts — Method

load_counts(data::DataMatrix{<:LazyMergedMatrix})

Merge/load counts for a lazily merged DataMatrix.

SingleCellProjections.loadh5ad — Method

loadh5ad(filename; var_id_column=:id, obs_id_column=:id)

Experimental loading of .h5ad files.

SingleCellProjections.logtransform — Method

logtransform(counts::DataMatrix; scale_factor=10_000, var=:copy, obs=:copy)

Log-transform counts using the formula:

  log(1 + cᵢⱼ*scale_factor/(∑ᵢcᵢⱼ))

var - Can be :copy (make a copy of source var) or :keep (share the source var object).

obs - Can be :copy (make a copy of source obs) or :keep (share the source obs object).

Examples

julia> transformed = logtransform(counts)

See also: sctransform

SingleCellProjections.merge_counts — Method

merge_counts(samples, sample_names;
             lazy=false,
             var_id_cols=nothing,
             sample_name_col = sample_names===nothing ? nothing : "sampleName",
             merged_obs_id_col = "id",
             merged_obs_id_delim = '_',
             callback=nothing)

Merge samples to create one large DataMatrix, by concatenating the obs. The union of the variables from the samples is used, and if a variable is not present in a sample, the count will be set to zero.

The obs IDs are created by concatenating the current obs ID columns, together with the sample_names (if provided).

lazy - Lazy merging. Use load_counts to actually perform the merging.

var_id_cols - Set to nothing to autodetect from the samples.

sample_name_col - Column in which the sample_names are stored.

merged_obs_id_col - obs ID column after merging. Set to nothing to keep current ID cols (will add sampleName as a separate ID column, if provided.)

merged_obs_id_delim - Delimiter used when merging obs IDs.

callback - Experimental callback functionality. The callback function is called between samples during merging. Return true to abort loading and false to continue.

See also: load_counts

SingleCellProjections.normalize_matrix — Method

normalize_matrix(data::DataMatrix, design::DesignMatrix; scale=false, kwargs...)

Normalize data using the specified design matrix.

See also: NormalizationModel, designmatrix

SingleCellProjections.normalize_matrix — Method

normalize_matrix(data::DataMatrix, [covariates...]; center=true, scale=false, kwargs...)

Normalize data. By default, the matrix is centered. Any covariates specified (using column names of data.obs) will be regressed out.

center - Set to true to center the data matrix.

scale - Set to true to scale the variables in the data matrix to unit standard deviation.

For other kwargs and more detailed descriptions, see NormalizationModel and designmatrix.

Examples

Centering only:

julia> normalize_matrix(data)

Regression model with intercept (centering) and "fraction_mt" (numerical annotation):

julia> normalize_matrix(data, "fraction_mt")

As above, but also including "batch" (categorical annotation):

julia> normalize_matrix(data, "fraction_mt", "batch")

See also: NormalizationModel, designmatrix

SingleCellProjections.obs_coordinates — Function

obs_coordinates(data::DataMatrix)

Returns a matrix with coordinates for the observations. Not available for all types of DataMatrices. Mostly useful for data matrices after dimension reduction such as svd or force_layout has been applied.

In the case of SVD (PCA), obs_coordinates returns the principal components, scaled by the singular values. This is a a good starting point for downstream analysis, since it is the optimal linear approximation of the original data for the given number of dimensions.

SingleCellProjections.project — Method

project(data::DataMatrix, model::ProjectionModel, args...; verbose=true, kwargs...)

Convenience function for projection onto multiple models. Essentially calls foldl and prints some @info messages (if verbose=true). In most cases, it is better to call project(data, base::DataMatrix) instead of using this method directly.

SingleCellProjections.project — Method

project(data::DataMatrix, base::DataMatrix, args...; from=nothing, kwargs...)

Project data onto base, by applying ProjectionModels from base one by one.

Since data already might have some models applied, project will try to figure out which models from base to use. See "Examples" below for concrete examples. Here's a more technical overview:

Consider a base data matrix with four models:

base: A -> B -> C -> D

Given some new data (typically counts), we can project that onto base, given the result proj by applying all four models:

data:
proj: A -> B -> C -> D

If data already has some models applied (e.g. we already projected onto A and B above), project will look for the last model in data (in this case B) in the list of models in base, and only apply models after that (in this case C and D).

data: A -> B
proj: A -> B -> C -> D

It is also possible to use the from kwarg to specify exactly which models to apply. (The models in from must be a prefix of the models in base, or in other words, base was created by applying additional operations to from.)

data: X
base: A -> B -> C -> D
from: A -> B
proj: X -> C -> D

Note that it is necessary to use the from kwarg if the last model in data does not occurr in base, because project cannot figure out on its own which models it makes sense to apply.

Examples

First, we construct a "base" by loading counts, SCTransforming, normalizing, computing the svd and finally computing a force layout:

julia> fp = ["GSE164378_RNA_ADT_3P_P1.h5", "GSE164378_RNA_ADT_3P_P2.h5"];
julia> counts = load_counts(fp; sample_names=["P1","P2"]);
julia> transformed = sctransform(counts);
julia> normalized = normalize_matrix(transformed);
julia> reduced = svd(normalized; nsv=10);
julia> fl = force_layout(reduced; ndim=3, k=100)
  DataMatrix (3 variables and 35340 observations)
  Matrix{Float64}
  Variables: id
  Observations: id, sampleName, barcode
  Models: NearestNeighborModel(base="force_layout", k=10), SVD, Normalization, SCTransform

Note how the last line lists all ProjectionModels used in the creation of fl.

Next, let's load some more samples for projection:

julia> fp2 = ["GSE164378_RNA_ADT_3P_P5.h5", "GSE164378_RNA_ADT_3P_P6.h5"];
julia> counts2 = load_counts(fp2; sample_names=["P5","P6"]);

It is easy to project the newly loaded counts2 onto the "base" force layout fl:

julia> project(counts2, fl)
DataMatrix (3 variables and 42553 observations)
  Matrix{Float64}
  Variables: id
  Observations: id, sampleName, barcode
  Models: NearestNeighborModel(base="force_layout", k=10), SVD, Normalization, SCTransform

We can also project in two or more steps, to get access to intermediate results:

julia> reduced2 = project(counts2, reduced)
DataMatrix (20239 variables and 42553 observations)
  SVD (10 dimensions)
  Variables: id, feature_type, name, genome, read, pattern, sequence, logGeneMean, outlier, beta0, ...
  Observations: id, sampleName, barcode
  Models: SVDModel(nsv=10), Normalization, SCTransform

julia> project(reduced2, fl)
DataMatrix (3 variables and 42553 observations)
  Matrix{Float64}
  Variables: id
  Observations: id, sampleName, barcode
  Models: NearestNeighborModel(base="force_layout", k=10), SVD, Normalization, SCTransform

If the DataMatrix we want to project is modified, we need to use the from kwarg to tell project which models to use:

julia> filtered = counts2[:,1:10:end]
DataMatrix (33766 variables and 4256 observations)
  SparseArrays.SparseMatrixCSC{Int64, Int32}
  Variables: id, feature_type, name, genome, read, pattern, sequence
  Observations: id, sampleName, barcode
  Models: FilterModel(:, 1:10:42551)

julia> reduced2b = project(filtered2, reduced; from=counts)
DataMatrix (20239 variables and 4256 observations)
  SVD (10 dimensions)
  Variables: id, feature_type, name, genome, read, pattern, sequence, logGeneMean, outlier, beta0, ...
  Observations: id, sampleName, barcode
  Models: SVDModel(nsv=10), Normalization, SCTransform, Filter

After that, it is possible to continue without specifying from:

julia> project(reduced2b, fl)
DataMatrix (3 variables and 4256 observations)
  Matrix{Float64}
  Variables: id
  Observations: id, sampleName, barcode
  Models: NearestNeighborModel(base="force_layout", k=10), SVD, Normalization, SCTransform, Filter

SingleCellProjections.project — Method

project(data::DataMatrix, model::ProjectionModel, args...; verbose=true, kwargs...)

Core projection function. Project data based on the single ProjectionModel model. In most cases, it is better to call project(data, base::DataMatrix) instead of using this method directly.

SingleCellProjections.splitrange — Method

splitrange(r::UnitRange, nparts::Integer)

Splits a range in nparts number of parts of equal length.

SingleCellProjections.tf_idf_transform — Method

tf_idf_transform(counts::DataMatrix;
                 scale_factor = 10_000,
                 idf = vec(size(counts,2) ./ max.(1,sum(counts.matrix; dims=2))),
                 annotate = true,
                 var = :copy,
                 obs = :copy)

Compute the TF-IDF (term frequency-inverse document frequency) transform of counts, using the formula log( 1 + scale_factor * tf * idf ) where tf is the term frequency counts.matrix ./ max.(1, sum(counts.matrix; dims=1)).

If annotate is true, idf will be added as a var annotation.

SingleCellProjections.update_matrix — Function

update_matrix(data::DataMatrix, matrix, model=nothing;
              var::Union{Symbol,String,DataFrame} = "",
              obs::Union{Symbol,String,DataFrame} = "",
              var_id_cols,
              obs_id_cols)

Create a new DataMatrix by replacing parts of data with new values. Mostly useful when implementing new ProjectionModels.

matrix - the new matrix.

model - will be appended to the list of models from data. If set to nothing, the resulting list of models will be empty.

Kwargs:

var - One of:

:copy - Copy from data.

:keep - Share var with data.

::DataFrame - Replace with a new table with variable annotations.

prefix::String - Prefix, the new variables will be named prefix1, prefix2, etc.

obs See var.

var_id_cols - New ID columns. Defaults to the same as data, or "id" if new variables were generated using the "prefix" above.

obs_id_cols - See var_id_cols.

SingleCellProjections.var_coordinates — Function

var_coordinates(data::DataMatrix)

Returns a matrix with coordinates for the variables. Only available for DataMatrices that have a dual representation (e.g. SVD/PCA).

In the case of SVD (PCA), var_coordinates returns the principal components as unit vectors.

SingleCellProjections.var_counts_fraction! — Method

var_counts_fraction!(counts::DataMatrix, sub_filter, tot_filter, col; check=true)

For each observation, compute the fraction of counts that match a specific variable pattern.

sub_filter decides which variables are counted.

tot_filter decides which variables to include in the total.

If check=true, an error will be thrown if no variables match the patterns.

For more information on filtering syntax, see examples below and the documentation on DataFrames.filter.

Examples

Compute the fraction of reads in MT- genes, considering only "Gene Expression" features (and not e.g. "Antibody Capture").

var_counts_fraction!(counts, "name"=>contains(r"^MT-"), "feature_type"=>isequal("Gene Expression"), "fraction_mt")

Compute the fraction of reads in MT- genes, when there is no feature_type annotation (i.e. all variables are genes).

var_counts_fraction!(counts, "name"=>contains(r"^MT-"), Returns(true), "fraction_mt")

SingleCellProjections.variable_std — Method

variable_std(data::DataMatrix)

Computes the variance of each variable in data.

Note

data must be mean-centered. E.g. by using normalize_matrix before calling variable_std.

See also: variable_var, normalize_matrix

SingleCellProjections.variable_var — Method

variable_var(data::DataMatrix)

Computes the variance of each variable in data.

Note

data must be mean-centered. E.g. by using normalize_matrix before calling variable_var.

See also: variable_std, normalize_matrix